rename and rearrange some portfolio route functions
