+++ templates/portfolios/admin.html
@@ -17,7 +17,7 @@
 
         {% if user_can(permissions.VIEW_PORTFOLIO_NAME) %}
         <base-form inline-template>
-          <form method="POST" action="{{ url_for('portfolios.edit_portfolio', portfolio_id=portfolio.id) }}" autocomplete="false">
+          <form method="POST" action="{{ url_for('portfolios.edit', portfolio_id=portfolio.id) }}" autocomplete="false">
             {{ portfolio_form.csrf_token }}
             <div class='form-row'>
               <div class='form-col form-col--half'>
@@ -54,7 +54,7 @@
 
     {% if user_can(permissions.VIEW_PORTFOLIO_ACTIVITY_LOG) %}
       {% include "fragments/audit_events_log.html" %}
-      {{ Pagination(audit_events, 'portfolios.portfolio_admin', portfolio_id=portfolio.id) }}
+      {{ Pagination(audit_events, 'portfolios.admin', portfolio_id=portfolio.id) }}
     {% endif %}
   </div>
 {% endblock %}
