+++ templates/portfolios/header.html
@@ -60,8 +60,8 @@
       {{ Link(
         icon='chart-pie',
         text='navigation.portfolio_navigation.breadcrumbs.reports' | translate,
-        url=url_for("portfolios.portfolio_reports", portfolio_id=portfolio.id),
-        active=request.url_rule.endpoint == "portfolios.portfolio_reports",
+        url=url_for("portfolios.reports", portfolio_id=portfolio.id),
+        active=request.url_rule.endpoint == "portfolios.reports",
       ) }}
     {% endif %}
     {{ Link(
@@ -74,8 +74,8 @@
       {{ Link(
         icon='cog',
         text='navigation.portfolio_navigation.breadcrumbs.admin' | translate,
-        url=url_for("portfolios.portfolio_admin", portfolio_id=portfolio.id),
-        active=request.url_rule.endpoint == "portfolios.portfolio_admin",
+        url=url_for("portfolios.admin", portfolio_id=portfolio.id),
+        active=request.url_rule.endpoint == "portfolios.admin",
       ) }}
     {% endif %}
   </div>
